requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
click==8.1.7
distro==1.9.0
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.6.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.32.0.post1
tqdm==4.66.1
typing_extensions==4.9.0
uvicorn==0.25.0
